# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/stores":{"get":{"operationId":"StoreController_getAllStore","summary":"모든 스토어 정보 찾기","parameters":[{"name":"title","required":true,"in":"query","example":"홍대 팝업 스토어","description":"팝업 스토어 이름","schema":{"type":"string"}},{"name":"description","required":true,"in":"query","example":"홍대 팝업 스토어 오픈! ...","description":"팝업 스토어 설명","schema":{"type":"string"}},{"name":"category","required":true,"in":"query","example":"의류","description":"팝업 스토어 종류","schema":{"type":"string"}},{"name":"start_date","required":true,"in":"query","example":"2023-05-31T12:34:56.789Z","description":"팝업 스토어 시작일","schema":{"format":"date-time","type":"string"}},{"name":"end_date","required":true,"in":"query","example":"2023-05-31T12:34:56.789Z","description":"팝업 스토어 종료일","schema":{"format":"date-time","type":"string"}},{"name":"location","required":true,"in":"query","example":"서울시 성동구 OOO","description":"팝업 스토어 위치","schema":{"type":"string"}},{"example":"대구","description":"스토어 위치(시)","required":true,"name":"sido","in":"query","schema":{"type":"string"}},{"example":"중구","description":"스토어 위치(군/구)","required":true,"name":"sigungu","in":"query","schema":{"type":"string"}},{"example":"Point","description":"스토어 위치 검색용(Point 고정)","required":true,"name":"type","in":"query","schema":{"type":"string"}},{"example":"[127.23414, 34.1234]","description":"스토어 위치 검색용(좌표)","required":true,"name":"coordinates","in":"query","schema":{"type":"array"}},{"name":"price","required":true,"in":"query","example":"3000","description":"팝업 스토어 입장료","schema":{"type":"number"}},{"name":"sns","required":true,"in":"query","example":"[{insta, 홍대 인스타그램, http://...}]","description":"팝업 스토어 SNS 정보","schema":{"type":"array","items":{"type":"string"}}},{"name":"reservation_required","required":true,"in":"query","example":"true","description":"팝업 스토어 예약 필요 여부","schema":{"type":"boolean"}},{"name":"images","required":true,"in":"query","example":"[{1.png, 2.png}]","description":"팝업 스토어 이미지","schema":{"type":"array","items":{"type":"string"}}},{"name":"scrap","required":true,"in":"query","example":"[123, 424, 124]","description":"팝업 스토어 스크랩 리스트","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":""}},"tags":["Store"]},"post":{"operationId":"StoreController_create","summary":"스토어 정보 등록하기","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreRequestDto"}}}},"responses":{"201":{"description":""}},"tags":["Store"]}},"/stores/store/{id}":{"get":{"operationId":"StoreController_getStoreById","summary":"ID로 스토어 정보 찾기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Store"]}},"/stores/coord":{"get":{"operationId":"StoreController_getStoresByCoord","summary":"좌표 위치 주변 스토어 정보 찾기","parameters":[],"responses":{"200":{"description":""}},"tags":["Store"]}},"/stores/{id}":{"put":{"operationId":"StoreController_update","summary":"스토어 정보 수정하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StoreRequestDto"}}}},"responses":{"200":{"description":""}},"tags":["Store"]},"delete":{"operationId":"StoreController_delete","summary":"스토어 정보 삭제하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Store"]}},"/users":{"get":{"operationId":"UserController_getAllUsers","summary":"모든 유저 정보 찾기","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]},"post":{"operationId":"UserController_createUser","summary":"유저 정보 등록하기","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignupDto"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/{id}":{"get":{"operationId":"UserController_getUserById","summary":"ID로 유저 정보 찾기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]},"patch":{"operationId":"UserController_updateUser","summary":"유저 정보 수정하기(token 필요)","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["User"],"security":[{"Authorization":[]}]},"delete":{"operationId":"UserController_deleteUser","summary":"유저 정보 삭제하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/users/checknickname":{"post":{"operationId":"UserController_checkNickname","summary":"유저 닉네임 중복 여부 확인하기","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/checknickname"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/checkemail":{"post":{"operationId":"UserController_checkEmail","summary":"유저 이메일 중복 여부 확인하기","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/checkemail"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/{userId}/scrapStore/{storeId}":{"patch":{"operationId":"UserController_updateScrap","summary":"유저 스크랩 & 스토어 스크랩 등록하기","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"storeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/users/{userId}/unscrapStore/{storeId}":{"patch":{"operationId":"UserController_updateUnScrap","summary":"유저 스크랩 & 스토어 스크랩 취소하기","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"storeId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/posts":{"get":{"operationId":"PostsController_getAllPosts","summary":"모든 게시글 찾기","parameters":[],"responses":{"200":{"description":""}},"tags":["Post"]},"post":{"operationId":"PostsController_createPost","summary":"게시글 등록하기","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["Post"]}},"/posts/gather":{"get":{"operationId":"PostsController_getAllGatherPosts","summary":"모집게시판 게시글 찾기","parameters":[],"responses":{"200":{"description":""}},"tags":["Post"]}},"/posts/review":{"get":{"operationId":"PostsController_getAllReviewPosts","summary":"후기게시판 게시글 찾기","parameters":[],"responses":{"200":{"description":""}},"tags":["Post"]}},"/posts/free":{"get":{"operationId":"PostsController_getAllFreePosts","summary":"자유게시판 게시글 찾기","parameters":[],"responses":{"200":{"description":""}},"tags":["Post"]}},"/posts/{id}":{"get":{"operationId":"PostsController_getPostById","summary":"ID로 게시글 찾기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Post"]},"patch":{"operationId":"PostsController_updatePost","summary":"게시글 수정하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["Post"]},"delete":{"operationId":"PostsController_deletePost","summary":"게시글 삭제하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Post"]}},"/posts/{id}/views":{"patch":{"operationId":"PostsController_incrementViewCount","summary":"게시글 조회수 증가하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Post"]}},"/notifications":{"get":{"operationId":"NotificationsController_getNotifications","summary":"모든 알림 조회","parameters":[],"responses":{"200":{"description":""}},"tags":["Notification"]},"post":{"operationId":"NotificationsController_createNotification","summary":"알림 등록하기","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["Notification"]}},"/notifications/user/{userId}":{"get":{"operationId":"NotificationsController_getUserNotifications","summary":"유저별 알림 조회","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notification"]}},"/notifications/{id}":{"get":{"operationId":"NotificationsController_getNotificationsById","summary":"ID로 알림 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notification"]},"patch":{"operationId":"NotificationsController_updateNotification","summary":"알림 수정하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["Notification"]},"delete":{"operationId":"NotificationsController_deleteNotification","summary":"알림 삭제하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notification"]}},"/comments":{"get":{"operationId":"CommentsController_getAllComments","summary":"모든 댓글 조회","parameters":[],"responses":{"200":{"description":""}},"tags":["Comment"]},"post":{"operationId":"CommentsController_createPost","summary":"댓글 추가하기","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentCreateDto"}}}},"responses":{"201":{"description":""}},"tags":["Comment"]}},"/comments/{id}":{"get":{"operationId":"CommentsController_getCommentById","summary":"ID로 댓글 조회","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Comment"]},"patch":{"operationId":"CommentsController_updateComment","summary":"댓글 수정하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentUpdateDto"}}}},"responses":{"200":{"description":""}},"tags":["Comment"]},"delete":{"operationId":"CommentsController_deleteComment","summary":"댓글 삭제하기","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Comment"]}},"/auth/login":{"post":{"operationId":"AuthController_logIn","summary":"유저 로그인 토큰 발급","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","summary":"유저 로그인 토큰 인증","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"],"security":[{"Authorization":[]}]}}},"info":{"title":"POPular API","description":"POPular API입니다.","version":"4.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"Authorization","in":"header"}},"schemas":{"day":{"type":"object","properties":{}},"postinfo":{"type":"object","properties":{"sido":{"type":"string","example":"대구","description":"스토어 위치(시)"},"sigungu":{"type":"string","example":"중구","description":"스토어 위치(군/구)"}},"required":["sido","sigungu"]},"coordinfo":{"type":"object","properties":{"type":{"type":"string","example":"Point","description":"스토어 위치 검색용(Point 고정)"},"coordinates":{"example":"[127.23414, 34.1234]","description":"스토어 위치 검색용(좌표)","type":"array","items":{"type":"string"}}},"required":["type","coordinates"]},"StoreRequestDto":{"type":"object","properties":{"title":{"type":"string","example":"홍대 팝업 스토어","description":"팝업 스토어 이름"},"description":{"type":"string","example":"홍대 팝업 스토어 오픈! ...","description":"팝업 스토어 설명"},"category":{"type":"string","example":"의류","description":"팝업 스토어 종류"},"start_date":{"format":"date-time","type":"string","example":"2023-05-31T12:34:56.789Z","description":"팝업 스토어 시작일"},"end_date":{"format":"date-time","type":"string","example":"2023-05-31T12:34:56.789Z","description":"팝업 스토어 종료일"},"hours":{"example":"{{mon: {08:00, 18:00} ... }}","description":"팝업 스토어 운영시간","allOf":[{"$ref":"#/components/schemas/day"}]},"location":{"type":"string","example":"서울시 성동구 OOO","description":"팝업 스토어 위치"},"postcode":{"example":"{ sido: 대구, sigungu: 중구 }","description":"팝업 스토어 시, 군/구","allOf":[{"$ref":"#/components/schemas/postinfo"}]},"coord":{"example":"{ type: Point, coordinates: [127.6346, 34.1245](경도, 위도 순) }","description":"팝업 스토어 위치 좌표","allOf":[{"$ref":"#/components/schemas/coordinfo"}]},"price":{"type":"number","example":"3000","description":"팝업 스토어 입장료"},"sns":{"example":"[{insta, 홍대 인스타그램, http://...}]","description":"팝업 스토어 SNS 정보","type":"array","items":{"type":"string"}},"reservation_required":{"type":"boolean","example":"true","description":"팝업 스토어 예약 필요 여부"},"images":{"example":"[{1.png, 2.png}]","description":"팝업 스토어 이미지","type":"array","items":{"type":"string"}},"scrap":{"example":"[123, 424, 124]","description":"팝업 스토어 스크랩 리스트","type":"array","items":{"type":"string"}}},"required":["title","description","category","start_date","end_date","hours","location","postcode","coord","price","sns","reservation_required","images","scrap"]},"checknickname":{"type":"object","properties":{"nickname":{"type":"string","example":"귀여운 토끼","description":"중복 체크할 닉네임"}},"required":["nickname"]},"checkemail":{"type":"object","properties":{"email":{"type":"string","example":"elice@elice.com","description":"중복 체크할 이메일"}},"required":["email"]},"UserSignupDto":{"type":"object","properties":{"email":{"type":"string","example":"test1@elice.com","description":"유저 이메일"},"pw":{"type":"string","example":"1111","description":"유저 비밀번호"},"name":{"type":"string","example":"엘리스","description":"유저 이름"},"nickname":{"type":"string","example":"귀여운 토끼","description":"유저 닉네임"},"phone_number":{"type":"string","example":"010-1234-4321","description":"유저 전화번호"},"allow_notification":{"type":"boolean","example":"true","description":"유저 알림 여부"}},"required":["email","pw","name","nickname","phone_number","allow_notification"]},"UserUpdateDto":{"type":"object","properties":{"profile":{"type":"string","example":"1.png","description":"유저 프로필 사진"},"introduce":{"type":"string","example":"스포츠와 커피를 좋아하는 엘리스입니다.","description":"유저 한 줄 소개"},"nickname":{"type":"string","example":"귀여운 토끼","description":"유저 닉네임"},"pw":{"type":"string","example":"qwer1234","description":"유저 비밀번호"},"phone_number":{"type":"string","example":"010-1234-4321","description":"유저 전화번호"},"interested_category":{"example":"[의류, 식품]","description":"유저 선호 카테고리 목록","type":"array","items":{"type":"string"}},"allow_notification":{"type":"boolean","example":"true","description":"유저 알림 허용 여부"}},"required":["profile","introduce","nickname","pw","phone_number","interested_category","allow_notification"]},"ObjectId":{"type":"object","properties":{}},"PostCreateDto":{"type":"object","properties":{"title":{"type":"string","example":"홍대 팝업스토어 같이 가실 분?","description":"게시글 제목"},"author":{"example":"귀여운 토끼","description":"게시글 작성자","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"board":{"type":"string","example":"Gather","description":"게시글 올릴 게시판 종류"},"content":{"type":"string","example":"6월 16일 같이 가실 분 구합니다!","description":"게시글 내용"},"storeId":{"example":"qwerra21341","description":"게시글에서 설정한 스토어 ID","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"ratings":{"type":"number","example":"3.5","description":"게시글 평점"},"images":{"example":"[1.png, 2.png]","description":"게시글 이미지","type":"array","items":{"type":"string"}},"likes":{"example":"[qwer1234, rewq431]","description":"게시글 좋아요 누른 유저 목록","type":"array","items":{"type":"string"}},"reports":{"example":"[qwer1234, rewq431]","description":"게시글 신고 누른 유저 목록","type":"array","items":{"type":"string"}},"comments":{"example":"[저요!, 까비요]","description":"게시글 댓글","type":"array","items":{"type":"string"}},"views":{"type":"number","example":"4","description":"게시글 조회수"}},"required":["title","author","board","content","storeId","ratings","images","likes","reports","comments","views"]},"PostUpdateDto":{"type":"object","properties":{"title":{"type":"string","example":"홍대 팝업스토어 같이 가실 분?","description":"게시글 제목"},"content":{"type":"string","example":"6월 16일 같이 가실 분 구합니다!","description":"게시글 내용"},"images":{"example":"[1.png, 2.png]","description":"게시글 이미지","type":"array","items":{"type":"string"}},"storeId":{"type":"string","example":"qwerra21341","description":"게시글에서 설정한 스토어 ID"},"ratings":{"type":"number","example":"3.5","description":"게시글 평점"}},"required":["title","content","images","storeId","ratings"]},"NotificationCreateDto":{"type":"object","properties":{"type":{"type":"string","example":"Follow","description":"알림 종류"},"board":{"type":"string","example":"Gather","description":"알림이 온 게시판 종류"},"userId":{"example":"qwer2134","description":"알림 대상 ID","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"content":{"example":"귀여운 토끼님이 팔로우하셨습니다.","description":"알림 내용","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"contentModel":{"type":"string","example":"User","description":"내용 모델"}},"required":["type","board","userId","content","contentModel"]},"NotificationUpdateDto":{"type":"object","properties":{"checked":{"type":"boolean","example":"true","description":"알림 확인 여부(true: 안 읽음, false: 읽음)"}},"required":["checked"]},"CommentCreateDto":{"type":"object","properties":{"author":{"example":"qqwe2134","description":"댓글 작성자 ID","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"content":{"type":"string","example":"저요!","description":"댓글 내용"},"recomments":{"example":"저도 갈래요","description":"대댓글","type":"array","items":{"type":"string"}}},"required":["author","content","recomments"]},"CommentUpdateDto":{"type":"object","properties":{"content":{"type":"string","example":"저요!","description":"댓글 내용"}},"required":["content"]},"UserLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"test1@elice.com","description":"유저 이메일"},"pw":{"type":"string","example":"1111","description":"유저 비밀번호"}},"required":["email","pw"]}}}}